requirements.txt
streamlit
pandas
plotly
openai
python-dotenv 